base: move processing of the direct room inside the `AccountDataProcessor`
